requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
scipy==1.14.1
numpy==2.1.2
python-dotenv==1.0.1
apscheduler==3.10.4
alembic==1.13.3
pydantic==2.9.2
aiohttp==3.10.10
